removed subText.jsx, moved header.jsx out into /src/components, cleaned up styles.css for header
