+++ aunt-leahs-app/src/components/header.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import SubText from './SubText.jsx';
-import constants from '../../constants.jsx';
-import '../../styles.css';
+import constants from '../constants.jsx';
+import '../styles.css';
 
 const headerText = constants.HEADER.MAIN_HEADER;
 
@@ -9,10 +8,10 @@ const Head = ({ page }) => {
     /*takes in 'page' as props. Refer to ./components/constants.jsx for props.page values */
     return(
         <div className='headerBox'>
-                <img className="headerImage header"  alt="Aunt Leah's Logo" src={require('../../AuntLeahsTrees.png')}/>
+                <img className="headerImage header"  alt="Aunt Leah's Logo" src={require('../AuntLeahsTrees.png')}/>
                 <div className="headerMain header">{headerText.headTwo}</div>
                 {(page !== null) 
-                    &&<SubText title={page}/>}
+                    && <div className="headerSub header">{page}</div>}
         </div>
     )
 };
